requirements.txt
flask
pytesseract
pillow
flask-cors
pdfplumber
pdf2image